+++ src/misc.c
@@ -9,7 +9,7 @@
 #include "rand.h"
 
 /* Function to return the maximum of two variables */
-long double maximum(long double a, long double b)
+double maximum(double a, double b)
 {
     if (a > b) {
         return (a);
@@ -18,7 +18,7 @@ long double maximum(long double a, long double b)
 }
 
 /* Function to return the minimum of two variables */
-long double minimum(long double a, long double b)
+double minimum(double a, double b)
 {
     if (a < b) {
         return (a);
@@ -27,10 +27,10 @@ long double minimum(long double a, long double b)
 }
 
 /* Function to return the modulus of a vector */
-long double modulus(long double *x, int n)
+double modulus(double *x, int n)
 {
     int i;
-    long double res;
+    double res;
     res = 0.0;
     for (i = 0; i < n; i++) {
         res += x[i] * x[i];
@@ -39,10 +39,10 @@ long double modulus(long double *x, int n)
 }
 
 /* Function to return the dot product of two vecors */
-long double dot(long double *a, long double *b, int n)
+double dot(double *a, double *b, int n)
 {
     int i;
-    long double res;
+    double res;
     res = 0.0;
     for (i = 0; i < n; i++) {
         res += a[i] * b[i];
@@ -51,13 +51,13 @@ long double dot(long double *a, long double *b, int n)
 }
 
 /* Function to return the mean of n variables */
-long double mean(long double *x, int n)
+double mean(double *x, int n)
 {
     int i;
-    long double res;
+    double res;
     res = 0.0;
     for (i = 0; i < n; i++) {
         res += x[i];
     }
-    return (res / (long double) n);
+    return (res / (double) n);
 }
